Services

+++ ZestFrontend/ViewModels/PostDetailsViewModel.cs
@@ -1,37 +1,36 @@
-﻿using CommunityToolkit.Maui.Views;
-using CommunityToolkit.Mvvm.ComponentModel;
+﻿using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
 using MauiIcons.Core;
 using Microsoft.AspNetCore.SignalR.Client;
-using System;
-using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.ComponentModel;
-using System.Linq;
-using System.Runtime.InteropServices;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows.Input;
 using ZestFrontend.DTOs;
+using ZestFrontend.Filters;
+using ZestFrontend.Pages;
 using ZestFrontend.Parameters;
 using ZestFrontend.Services;
 
 namespace ZestFrontend.ViewModels
 {
 	[QueryProperty(nameof(Post), "Post")]
-	public partial class PostDetailsViewModel : ObservableObject, INotifyPropertyChanged
+	public partial class PostDetailsViewModel : ObservableObject
 	{
 		AuthService _authService;
 		PostsService _postsService;
 		LikesService _likesService;
 		SignalRConnectionService _signalRConnectionService;
 		LikesHubConnectionService _likesHubConnectionService;
-		CommentsHubConnectionService _commentHubConnectionService;
+		DeleteHubConnectionService _deleteHubConnectionService;
 		CommentService _commentService;
+		CommunityService _communityService;
 		MediaService _mediaService;
+		CommentFilter _filter;
+		private Task TaskInit;
+		private bool _isNavigatingToCommentDetailPage;
 
-		
-		public PostDetailsViewModel( PostsService postsService, LikesService likesService, CommentService commentService, MediaService mediaService, LikesHubConnectionService likesHubConnectionService, CommentsHubConnectionService commentHubConnectionService, SignalRConnectionService signalRConnectionService, AuthService authService)
+
+		public PostDetailsViewModel(PostsService postsService, LikesService likesService, CommentService commentService, MediaService mediaService, LikesHubConnectionService likesHubConnectionService, DeleteHubConnectionService deleteHubConnectionService, SignalRConnectionService signalRConnectionService, AuthService authService, CommunityService communityService)
 		{
 			_authService = authService;
 			_postsService = postsService;
@@ -39,14 +38,21 @@ namespace ZestFrontend.ViewModels
 			_commentService = commentService;
 			_mediaService = mediaService;
 			_likesHubConnectionService=likesHubConnectionService;
-			_commentHubConnectionService=commentHubConnectionService;
+			_deleteHubConnectionService=deleteHubConnectionService;
 			_signalRConnectionService = signalRConnectionService;
-			_commentHubConnectionService.Init();
+			TaskInit = Init();
 			ReplyCommand = new ReplyCommand(ExecuteReplyCommand);
 			_ = new MauiIcon();
+			_filter = CommentFilter.All;
+			filterButtonText = "See trending";
+			_communityService=communityService;
 		}
-
-		public CarouselView Carousel {  get; set; }
+		private async Task Init()
+		{
+			_deleteHubConnectionService.DeleteConnection.On<int>("CommentDeleted", UpdateComment);
+			_likesHubConnectionService.LikesConnection.On<int>("CommentLiked", UpdateComment);
+		}
+		public CarouselView Carousel { get; set; }
 		public ICommand ReplyCommand { get; }
 		[ObservableProperty]
 		PostDTO post;
@@ -59,83 +65,160 @@ namespace ZestFrontend.ViewModels
 		bool isCarouselVisible;
 		[ObservableProperty]
 		bool isMediaPlayerVisible;
+		[ObservableProperty]
+		bool isRefreshing;
 		
+		[ObservableProperty]
+		string filterButtonText;
 		[ObservableProperty]
 		string replyText;
 		public ObservableCollection<PostResourcesDTO> Resources { get; private set; } = new();
 
 
 		[RelayCommand]
-		async Task DislikePostAsync()
-		{
-			await _likesService.Like( Post.Id, 0, false);
-		}
-		[RelayCommand]
 		async Task LikePostAsync()
 		{
-			await _likesService.Like(Post.Id, 0, true);
+			if (Post.Like == null)
+			{
 
+				await _likesService.Like(Post.Id, 0, true);
+			}
+			else if (Post.Like.Value == true)
+			{
+				await _likesService.RemoveLike(Post.Like.Id, Post.Id, 0);
+			}
+			else if (Post.Like.Value == false)
+			{
+				await _likesService.RemoveLike(Post.Like.Id, Post.Id, 0);
+				await _likesService.Like(Post.Id, 0, true);
+			}
+		}
+		[RelayCommand]
+		async Task DislikePostAsync()
+		{
+			if (Post.Like == null)
+			{
+
+				await _likesService.Like(Post.Id, 0, false);
+			}
+			else if (Post.Like.Value == false)
+			{
+				await _likesService.RemoveLike(Post.Like.Id, Post.Id, 0);
+			}
+			else if (Post.Like.Value == true)
+			{
+				await _likesService.RemoveLike(Post.Like.Id, Post.Id, 0);
+				await _likesService.Like(Post.Id, 0, false);
+			}
 		}
 		[RelayCommand]
 		async Task DeletePostAsync()
 		{
-			var response = await _postsService.DeletePost(Post.Id);
-			var content = await response.Content.ReadAsStringAsync();
-			UpdateComment(int.Parse(content));
+			await _postsService.DeletePost(Post.Id);
 		}
 
-		public async void GetComments()
+		public async Task GetComments()
 		{
-			Comments.Clear();
-			var comments = await _commentService.GetComments(_authService.Id, Post.Id);
+
+			DateTime lastDate = new DateTime();
+			if (Comments.Count==0)
+			{
+				lastDate = DateTime.Now;
+			}
+			else
+			{
+				lastDate = Comments.Last().PostedOn;
+			}
+			var comments = await _commentService.GetComments(Post.Id, lastDate, 5);
 			if (comments != null)
 			{
-				foreach (var comment in comments)
+				foreach (var comment in comments.Reverse())
 				{
-					comment.IsOwner = comment.Publisher==_authService.Username;
-					await IsOwner(comment.Replies);
+					if ((Post.IsModerator == true || comment.Publisher==_authService.Username || _authService.IsAdmin)  && comment.Publisher != "Unknown")
+					{
+						comment.IsOwner = true;
+					}
+					await IsOwner(comment.Replies, 0);
 					Comments.Add(comment);
+
 				}
 			}
+			_filter = CommentFilter.All;
+			FilterButtonText = "See trending";
+		}
+		public async Task GetTrendingComments()
+		{
+			
+			var skipIds = Comments.Select(x => x.Id).ToArray();
+
+			foreach (var comment in await _commentService.GetTrendingPostsAsync(50, Post.Id, skipIds))
+			{
+				if ((Post.IsModerator == true || comment.Publisher==_authService.Username || _authService.IsAdmin)  && comment.Publisher != "Unknown")
+				{
+					comment.IsOwner = true;
+				}
+				await IsOwner(comment.Replies, 0);
+				Comments.Add(comment);
+			}
+			_filter = CommentFilter.Trending;
+			FilterButtonText = "See all";
 
 		}
-		public async Task IsOwner(IEnumerable<CommentDTO> comments)
+		public async Task IsOwner(IEnumerable<CommentDTO> comments, int level)
 		{
 			foreach (var comment in comments)
 			{
-				
-				comment.IsOwner = comment.Publisher==_authService.Username;
-				if(comment.Replies == null)
+				if((Post.IsModerator == true || comment.Publisher==_authService.Username || _authService.IsAdmin)  && comment.Publisher != "Unknown")
+				{
+					comment.IsOwner = true;
+				}
+				comment.AreRepliesVisible = level<=4;
+				if (comment.Replies == null)
 				{
 					return;
 				}
-				await IsOwner(comment.Replies);
-				
+				await IsOwner(comment.Replies, level + 1);
+
 			}
 
 		}
 		[RelayCommand]
 		async Task SendAsync(string text)
 		{
-			
-		 var response = await _commentService.PostComment(Post.Id, text);
+
+			var response = await _commentService.PostComment(Post.Id, text);
 			var content = await response.Content.ReadAsStringAsync();
 			AddComment(int.Parse(content));
 
 		}
-		partial void OnPostChanged(PostDTO value)
+		async partial void OnPostChanged(PostDTO value)
 		{
-			GetComments();
-			DealWithResource();
+			Post.IsModerator = await _communityService.IsModerator(_authService.Id, Post.CommunityId);
+			if ((Post.Publisher == _authService.Username|| Post.IsModerator || _authService.IsAdmin) && Post.Publisher != "Unknown")
+			{
+				Post.IsOwner = true;
+			}
 			
-			_commentHubConnectionService.CommentsConnection.On("CommentPosted", GetComments);
-			_likesHubConnectionService.LikesConnection.On<int>("CommentLiked", UpdateComment);
+			Comments.Clear();
+			await GetComments();
+			DealWithResource();
+					
+			
 		}
-		
+
 		[RelayCommand]
 		async Task RefreshAsync()
 		{
-		    GetComments();
+			Comments.Clear();
+			if(_filter == CommentFilter.All) 
+			{
+				await GetComments();
+			}
+			else if(_filter == CommentFilter.Trending)
+			{
+				await GetTrendingComments();
+			}
+			IsRefreshing = false;
 		}
 		[RelayCommand]
 		async Task GoBackAsync()
@@ -145,69 +228,125 @@ namespace ZestFrontend.ViewModels
 		[RelayCommand]
 		async Task LikeCommentAsync(CommentDTO commentDTO)
 		{
-			await _likesService.Like(Post.Id, commentDTO.Id, true);
+			if (commentDTO.Like == null)
+			{
+
+				await _likesService.Like(Post.Id, commentDTO.Id, true);
+			}
+			else if (commentDTO.Like.Value == true)
+			{
+				await _likesService.RemoveLike(commentDTO.Like.Id, Post.Id, commentDTO.Id);
+			}
+			else if (commentDTO.Like.Value == false)
+			{
+				await _likesService.RemoveLike(commentDTO.Like.Id, Post.Id, commentDTO.Id);
+				await _likesService.Like(Post.Id, commentDTO.Id, true);
+			}
 		}
 		[RelayCommand]
 		async Task DislikeCommentAsync(CommentDTO commentDTO)
 		{
-			await _likesService.Like(Post.Id, commentDTO.Id, false);
+			if (commentDTO.Like == null)
+			{
+
+				await _likesService.Like(Post.Id, commentDTO.Id, false);
+			}
+			else if (commentDTO.Like.Value == false)
+			{
+				await _likesService.RemoveLike(commentDTO.Like.Id, Post.Id, commentDTO.Id);
+			}
+			else if (commentDTO.Like.Value == true)
+			{
+				await _likesService.RemoveLike(commentDTO.Like.Id, Post.Id, commentDTO.Id);
+				await _likesService.Like(Post.Id, commentDTO.Id, false);
+			}
 		}
 		[RelayCommand]
-		async Task ReplyCommentAsync(CommentDTO comment)
+		void ReplyCommentAsync(CommentDTO comment)
 		{
 			comment.IsReplyVisible = !comment.IsReplyVisible;
-			
+
 		}
-		
+
 		public async Task SendReplyAsync(int comment, string text)
 		{
 			var response = await _commentService.PostComment(Post.Id, text, comment);
 
-			
 			var content = await response.Content.ReadAsStringAsync();
 			string[] parts = content.Trim('[', ']').Split(',');
 			int firstNumber = int.Parse(parts[0]);
 			int secondNumber = int.Parse(parts[1]);
 			var reply = await _commentService.GetSingleComment(firstNumber);
-			reply.IsOwner = reply.Publisher == _authService.Username;
-			var commentToFind = FindCommentById(secondNumber, Comments);
-			commentToFind.Replies.Add(reply);
-
+			if(Post.IsModerator == true || reply.Publisher == _authService.Username)
+			{
+				reply.IsOwner = true;
+			}
+		
+			var commentToFind = FindCommentById(secondNumber, Comments, 0);
+			if (commentToFind != null && commentToFind.Length > 1)
+			{
+				var commentDto = (CommentDTO)commentToFind[0];
+				var commentLevel = (int)commentToFind[1];
+				commentDto.Replies.Add(reply);
+				if (commentLevel >= 4)
+				{
+					commentDto.AreRepliesVisible = false;
+				}
+			}
+			
 		}
-		public CommentDTO FindCommentById(int id, IEnumerable<CommentDTO> comments)
+		public object[] FindCommentById(int id, IEnumerable<CommentDTO> comments, int level)
 		{
 			foreach (var comment in comments)
 			{
 				if (comment.Id == id)
 				{
-					return comment; 
+					return new object[] { comment, level };
 				}
 
-				var foundInReplies = FindCommentById(id, comment.Replies);
+				var foundInReplies = FindCommentById(id, comment.Replies, level + 1);
 				if (foundInReplies != null)
 				{
-					return foundInReplies; 
+					return foundInReplies;
 				}
 			}
 
-			return null; 
+			return null;
 		}
+
 		public async void UpdateComment(int id)
 		{
 			var updatedComment = await _commentService.GetSingleComment(id);
-			var comment = FindCommentById(id, Comments);
-			comment.Likes = updatedComment.Likes;
-			comment.Dislikes = updatedComment.Dislikes;
+			var comment = FindCommentById(id, Comments, 0);
+
+			if (comment != null && comment.Length > 1)
+			{
+				var commentDto = (CommentDTO)comment[0];
+				commentDto.Likes = updatedComment.Likes;
+				commentDto.Dislikes = updatedComment.Dislikes;
+				commentDto.Like = updatedComment.Like;
+				commentDto.Text = updatedComment.Text;
+				commentDto.Publisher = updatedComment.Publisher;
+				if ((Post.IsModerator == true || commentDto.Publisher==_authService.Username || _authService.IsAdmin)  && commentDto.Publisher != "Unknown")
+				{
+					commentDto.IsOwner = true;
+				}
+			}
+
 		}
 		public async void AddComment(int id)
 		{
 			var comment = await _commentService.GetSingleComment(id);
+			if ((Post.IsModerator == true || comment.Publisher==_authService.Username || _authService.IsAdmin)  && comment.Publisher != "Unknown")
+			{
+				comment.IsOwner = true;
+			}
 			Comments.Add(comment);
 		}
 		[RelayCommand]
 		async Task DeleteCommentAsync(CommentDTO commentDTO)
 		{
-			await _commentService.DeleteComment(commentDTO.Id);
+			await _commentService.DeleteComment(commentDTO.Id, Post.Id);
 		}
 		public async void DealWithResource()
 		{
@@ -224,10 +363,10 @@ namespace ZestFrontend.ViewModels
 				IsMediaPlayerVisible = false;
 				PostResourcesDTO[] results = await _mediaService.GetPhotosByPostId(Post.Id);
 
-				Resources = new ObservableCollection<PostResourcesDTO>(results); // Convert the list to an ObservableCollection
-				
+				Resources = new ObservableCollection<PostResourcesDTO>(results);
+
 				OnPropertyChanged(nameof(Resources));
-				
+
 			}
 			else if (Post.ResourceType.Trim()=="video")
 			{
@@ -245,21 +384,89 @@ namespace ZestFrontend.ViewModels
 				return;
 			}
 		}
-
+		[RelayCommand]
+		async Task LoadMoreComments()
+		{
+			if (_filter == CommentFilter.All)
+				await GetComments();
+			else
+				await GetTrendingComments();
+		}
+		[RelayCommand]
+		async Task FilterCommentsAsync()
+		{
+			Comments.Clear();
+			if (_filter == CommentFilter.Trending)
+			{
+				await GetComments();
+			}
+			else
+			{
+				await GetTrendingComments();
+			}
+		}
+		
 		private async void ExecuteReplyCommand(ReplyCommandParameter parameter)
 		{
 			var comment = int.Parse(parameter.Comment);
 			var text = parameter.ReplyText;
 			await SendReplyAsync(comment, text);
-		}
-		public async void OnNavigatedTo()
-		{
+			var commentToFind = FindCommentById(comment, Comments, 0);
+			if (commentToFind != null && commentToFind.Length > 1)
+			{
+				var commentDto = (CommentDTO)commentToFind[0];
+				commentDto.IsReplyVisible = false;
+			}
 			
-			await _signalRConnectionService.RemoveConnectionToGroup(_likesHubConnectionService.LikesConnection.ConnectionId);
-			await _signalRConnectionService.RemoveConnectionToGroup(_commentHubConnectionService.CommentsConnection.ConnectionId);
-			await _signalRConnectionService.AddConnectionToGroup(_likesHubConnectionService.LikesConnection.ConnectionId, new string[] { $"pd-{Post.Id}", Post.Id.ToString() });
-			await _signalRConnectionService.AddConnectionToGroup(_commentHubConnectionService.CommentsConnection.ConnectionId, new string[] { $"message-{Post.Id}" });
+
 		}
-		
+		[RelayCommand]
+		async Task GoToCommentDetailPageAsync(CommentDTO comment)
+		{
+			if (comment == null) return;
+			_isNavigatingToCommentDetailPage = true;
+			await Shell.Current.GoToAsync($"{nameof(CommentDetailsPage)}?id={comment.Id}", true,
+				new Dictionary<string, object>
+			{
+			{"Comment", comment },
+			{"postId", Post }
+			});
+
+		}
+
+		public async Task OnNavigatedTo()
+		{
+			if (TaskInit is not null && !TaskInit.IsCompleted) await TaskInit;
+			if (!_isNavigatingToCommentDetailPage)
+			{
+				if(_likesHubConnectionService.LikesConnection.ConnectionId != null)
+				{
+					await _signalRConnectionService.AddConnectionToGroup(_likesHubConnectionService.LikesConnection.ConnectionId, new string[] { $"pdl-{Post.Id}", Post.Id.ToString() });
+				}
+				if(_deleteHubConnectionService.DeleteConnection.ConnectionId != null)
+				{
+					await _signalRConnectionService.AddConnectionToGroup(_deleteHubConnectionService.DeleteConnection.ConnectionId, new string[] { $"pdd-{Post.Id}" });
+				}
+				_authService.Groups.Add(Post.Id.ToString());
+				_authService.Groups.Add($"pdl-{Post.Id}");
+				_authService.Groups.Add($"pdd-{Post.Id}");
+			}
+			_isNavigatingToCommentDetailPage = false;
+		}
+		public async Task OnNavigatedFrom()
+		{
+			if (!_isNavigatingToCommentDetailPage)
+			{
+				if (_likesHubConnectionService.LikesConnection.ConnectionId != null)
+				{
+					await _signalRConnectionService.RemoveConnectionToGroup(_likesHubConnectionService.LikesConnection.ConnectionId);
+				}
+				if (_deleteHubConnectionService.DeleteConnection.ConnectionId != null)
+				{
+					await _signalRConnectionService.RemoveConnectionToGroup(_deleteHubConnectionService.DeleteConnection.ConnectionId);
+				}
+				_authService.Groups.Clear();
+			}
+		}
 	}
 }
